move keyboard accessory to it's own file
